requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
gunicorn>=20.1
psycopg2-binary>=2.9
python-dotenv>=1.0
whitenoise>=6.6
